Add support for go-nerve's labels
go-nerve is a rewrite of AirBnB's nerve for service discovery.
https://github.com/blablacar/go-nerve

You can add tags in the discovery process, that can be used for relabeling.
This is useful for SD to be able to push tags directly to prometheus.

This commit is backwards compatible with the original Nerve without the labels feature.